## Increase difficulty of game: 
1. Added a `difficultyMultiplier` that increases over time, capping at 3.0 after 2 minutes.
2. Reduced the enemy spawn interval from every 3 seconds to every 2 seconds.
3. Increased the spawn amount by changing the divisor from 30 to 20 and multiplying by the difficulty multiplier.
4. Reduced the delay between spawns in a group from 300ms to 200ms.
5. Made enemies move faster towards the player by changing the speed calculation divisor from 10.0 to 5.0.
6. Slightly increased the base enemy speed by changing it from 11.73 to 10.0 (lower number means faster movement).

These adjustments will result in:
- Enemies spawning more frequently
- More enemies spawning at once
- Enemies moving faster towards the player
- The difficulty ramping up more quickly over time